AnimalAdoptSystem/requirements.txt
Django==6.0.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.0
python-decouple==3.8
pytest-django==4.7.0
pytest-cov==4.1.0
requests==2.31.0
pyjwt==2.8.0
channels==4.0.0
channels-redis==4.3.0
redis==5.0.1
django-redis==6.0.0
celery==5.3.6
redis==5.0.1
